feat: Merge file assistant into Files page with unified client
Tab the existing Files (S3 workspace) page with a new Assistant tab,
removing the standalone /file-assistant route. Adds a typed
modelContextFilesService client for the OpenAI file-inputs proxy so
the dashboard dispatches workspace vs model-context uploads to the
right backend.
